CTF: Khronos Bugzilla #1671.
(7) Added functions to calculate the checksum of a file and the checksum of the entire test suite scripts. Added the calculation of the test file checksum and the test file post-processing script checksum, on execution. When exporting to HTML or CSV: write out the complete list of all the checksums.

(4) Added a judgement compiler so that the HTML and CSV exporters use the same code as the GUI to generate badge earned statements. Added the export of the judgement columns to the HTML and CSV exporters: no colors or fancy stuff though.

Also fixed serialization backward compatibility issue of judgement results and the COLLADA asset information.