CTF: Khronos Bugzilla #1671.
(9) Added the automatic creation of the "Default" settings. This setting is automatically selected when creating new test procedure.

(10) Fixed FJudgementContext::CompareImages: was returning the wrong result value. Added the 'tolerance' parameter to this function, defaults to 5.

(8) Fixed an exception when using the strangely named "Auto Update Result" button.

(7) Also calculate the checksums for the $CTF/ImageComparators python scripts...