Bugzilla Bug 1761.
First version of the partial refresh: works nicely in all my test procedures and the performance gain is appreciable. Modified the "auto-update executions" icon to a more friendly "refresh table" (although the code for it did not change). Modified all the scripts that forced a full refresh of the table to use the partial refresh on the tests that they affect.

There are three partial refresh functions. Use them this way:

1) For each 'test' to modify:
1.1) Call FExecutionGrid::PartialRefreshRemove(test).
1.2) Modify the test and its execution as necessary.
1.3) Call FExecutionGrid::PartialRefreshAdd(test).
2) Call FExecutionGrid::PartialRefreshDone().