Bugzilla Bug 1795.
Modified the COLLADA parser to read in the wanted asset elements: <keywords>, <subject> and <title>. I've renamed the Test Id into '#', because we still need this uniqueness information. Updated the execution grid and the CSV and HTML exporters to use the 'Categories', 'Description' and 'Test Id' information from the COLLADA document.