Bugzilla Bug 1841.
General blessing improvements: removed the image/animation distinction in the blessed viewer, added textbox to show the current default blessed image/animation, added a button to set the current default blessed image/animation. In the context menu: moved the blessing functions in their own sub-menu, "Bless [Image/Animation]", they are: "Replace Default", "Add as Default" and "Add as Alternate".

Reduced the padding in a couple of dialogs too, because the UI is absolutely huge right now in many dialogs.